+++ API/MetAPI/Next12HoursType.cs
@@ -1,8 +1,9 @@
-﻿using GraphQL.Types;
+﻿using Backend.API.MetAPI.SubClasses;
+using GraphQL.Types;
 
 namespace Backend.API.MetAPI
 {
-    public class Next12HoursType : ObjectGraphType<Forecast.Next12Hours>
+    public class Next12HoursType : ObjectGraphType<Next12Hours>
     {
         public Next12HoursType()
         {
