+++ API/MetAPI/SubClasses/Instant.cs
@@ -1,0 +1,9 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Backend.API.MetAPI.SubClasses
+{
+    public class Instant    {
+        [JsonPropertyName("details")]
+        public Details Details { get; set; } 
+    }
+}
