+++ API/MetAPI/SubClasses/Next12Hours.cs
@@ -1,0 +1,12 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Backend.API.MetAPI.SubClasses
+{
+    public class Next12Hours    {
+        [JsonPropertyName("summary")]
+        public Summary Summary { get; set; } 
+
+        [JsonPropertyName("details")]
+        public Details Details { get; set; } 
+    }
+}
